Isolate more create_integration helpers
